+++ lib/legion/transport/connection/ssl.rb
@@ -4,32 +4,40 @@ module Legion
   module Transport
     module Connection
       module SSL
-        def settings
+        def tls_options(tls_config: nil, port: nil)
+          return {} unless defined?(Legion::Crypt::TLS)
+
+          tls_config ||= tls_settings
+          port       ||= transport_port
+
+          tls = Legion::Crypt::TLS.resolve(tls_config, port: port)
+          return {} unless tls[:enabled]
+
+          {
+            tls:                 true,
+            tls_cert:            tls[:cert],
+            tls_key:             tls[:key],
+            tls_ca_certificates: [tls[:ca]].compact,
+            verify_peer:         tls[:verify] != :none
+          }
+        end
+
+        private
+
+        def tls_settings
+          return {} unless defined?(Legion::Settings)
+
           Legion::Settings[:transport][:tls] || {}
+        rescue StandardError
+          {}
         end
 
-        def use_vault_pki?
-          settings[:use_vault_pki] && Legion::Settings[:crypt][:vault][:connected]
-        end
+        def transport_port
+          return nil unless defined?(Legion::Settings)
 
-        def use_tls?
-          settings[:use_tls] || Legion::Settings[:transport][:connection][:port].to_i == 5671
-        end
-
-        def tls_cert
-          settings[:tls_cert]
-        end
-
-        def tls_key
-          settings[:tls_key]
-        end
-
-        def ca_certs
-          settings[:ca_certs]
-        end
-
-        def verify_peer?
-          settings[:verify_peer] || false
+          Legion::Settings[:transport][:connection][:port]
+        rescue StandardError
+          nil
         end
       end
     end